requirements.txt
Django==3.2.9
django-crontab==0.7.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
django-filter==22.1
